change data type of ndefined to int, because the corresponding API arguments ndims, nattrs, nvars are of type int, not MPI_Offset nor size_t.
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@3393 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ src/drivers/ncmpio/ncmpio_dim.c
@@ -218,7 +218,7 @@ ncmpio_def_dim(void       *ncdp,    /* IN:  NC object */
         }
     }
 
-    dimid = (int)ncp->dims.ndefined;
+    dimid = ncp->dims.ndefined;
 
     /* Add a new dim handle to the end of handle array */
     ncp->dims.value[dimid] = dimp;
